+++ byoao/src/vault/__tests__/obsidian-plugins.test.ts
@@ -26,6 +26,7 @@ function makePreset(
     name: "test",
     displayName: "Test",
     description: "test preset",
+    initOfferWhen: "always",
     directories: [],
     agentDescription: "test",
     frontmatterExtras: {},